Revert watching changes by previous touched status
